Better cast offset to long on ICC
git-svn-id: http://root.cern.ch/svn/root/trunk@27079 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/reflex/inc/Reflex/Builder/TypeBuilder.h
@@ -19,7 +19,7 @@
 #include <vector>
 
 #if defined(__ICC)
-#define OffsetOf(c1,mem) (int(&((volatile const char&)((c1*)0)->mem)))
+#define OffsetOf(c1,mem) (long(&((volatile const char&)((c1*)0)->mem)))
 #else
 #define OffsetOf(c1,mem) ((size_t)(&reinterpret_cast<const volatile char&>(((c1*)64)->mem))-64)
 #endif
